+++ src/components/input/search-icon-svg.tsx
